+++ src/bfregistry.rs
@@ -2,8 +2,9 @@ use std::{collections::HashMap, sync::Mutex};
 
 use once_cell::sync::Lazy;
 use tracing::info;
+use retour_utils::hook_module;
 
-use crate::{console::CommandError, debug_dll::get_from_memory};
+use crate::{console::{CommandError, add_to_command_register}, debug_dll::get_from_memory};
 
 static BF_REGISTRY: Lazy<Mutex<HashMap<String, u32>>> = Lazy::new(|| Mutex::new(HashMap::new()));
 
@@ -50,4 +51,38 @@ pub fn read_bf_registry() {
         }
         current_address += 0x1c;
     }
+}
+
+#[hook_module("zoo.exe")]
+mod zoo_bf_registry {
+    use crate::{
+        bfregistry::{add_to_registry, get_from_registry},
+        debug_dll::{get_from_memory, get_string_from_memory},
+    };
+
+    #[hook(unsafe extern "thiscall" BFRegistry_prtGetHook, offset = 0x000bdd22)]
+    fn prt_get(_this_prt: u32, class_name: u32, _delimeter_maybe: u8) -> u32 {
+        get_from_registry(get_string_from_memory(get_from_memory::<u32>(class_name))).unwrap()
+    }
+
+    #[hook(unsafe extern "cdecl" BFRegistry_AddHook, offset = 0x001770e5)]
+    fn add_to_bfregistry(param_1: u32, param_2: u32) -> u32 {
+        let param_1_string = get_string_from_memory(get_from_memory::<u32>(param_1));
+        add_to_registry(&param_1_string, param_2);
+        0x638001
+    }
+
+    #[hook(unsafe extern "cdecl" BFRegistry_AddUIHook, offset = 0x001774bf)]
+    fn add_to_bfregistry_ui(param_1: u32, param_2: u32) -> u32 {
+        let param_1_string = get_string_from_memory(get_from_memory::<u32>(param_1));
+        add_to_registry(&param_1_string, param_2);
+        0x638001
+    }
+}
+
+pub fn init() {
+    if let Err(e) = unsafe { zoo_bf_registry::init_detours() } {
+        info!("Error initialising bf_registry detours: {}", e);
+    };
+    add_to_command_register("list_bf_registry".to_owned(), command_list_registry)
 }
