+++ src/capture_ztlog.rs
@@ -1,4 +1,5 @@
 use tracing::info;
+use retour_utils::hook_module;
 
 enum ZTLogLevel {
     Trace,
@@ -36,4 +37,23 @@ impl ZTLogLevel {
 pub fn log_from_zt(source_file: &String, line_number: u32, level: u32, message: &String) {
     let level = ZTLogLevel::from_u32(level);
     info!("{}({}) : {} : {}", source_file, line_number, level.as_str(), message);
+}
+
+#[hook_module("zoo.exe")]
+mod zoo_logging {
+    use crate::{capture_ztlog::log_from_zt, debug_dll::get_string_from_memory};
+
+    #[hook(unsafe extern "cdecl" ZooLogging_LogHook, offset = 0x00001363)]
+    fn zoo_log_func(source_file: u32, param_2: u32, param_3: u32, _param_4: u8, _param_5: u32, _param_6: u32, log_message: u32) {
+        let source_file_string = get_string_from_memory(source_file);
+        let log_message_string = get_string_from_memory(log_message);
+        log_from_zt(&source_file_string, param_2, param_3, &log_message_string);
+    }
+}
+
+
+pub fn init() {
+    if let Err(e) = unsafe { zoo_logging::init_detours() } {
+        info!("Error initialising zt logging detours: {}", e);
+    };
 }
