+++ src/resource_manager.rs
